pyproject.toml
[tool.poetry]
name = "sample notification server"
version = "0.1.0"
description = ""
authors = ["Samuel Appau <appausamuel90@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.100.1"
uvicorn = "^0.23.1"
pre-commit = "^3.3.3"
pinject = "^0.14.1"
sqlalchemy = "^2.0.19"
redis = "^4.6.0"
gunicorn = "^21.2.0"
isort = "^5.12.0"
blinker = "^1.6.2"
alembic = "^1.11.1"
pytest = "^7.4.0"
python-dotenv = "^1.0.0"
pydantic = "^2.1.1"
fastapi-pagination = "^0.12.7"
pydantic-settings = "^2.0.2"
jwt = "^1.3.1"
psycopg2-binary = "^2.9.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
